Turbolinks is breaking things. Sorry turbolinks
When clicking on an "edit form" link in the workshop single
view, the subsequent form editing page was retrieved via
AJAX by turbolinks. This caused the JavaScript that loads the
form structure to not properly fires it's event handlers. Since
I did not much care to debug why, I resorted to simply removing
turbolinks.